Update handle authentication
