+++ ClearKeep/Scenes/Login/LoginRemoteStore.swift
@@ -9,10 +9,11 @@ import Foundation
 import Combine
 import ChatSecure
 import Model
+import Networking
 
 protocol ILoginRemoteStore {
 	func signIn(email: String, password: String, domain: String) async -> Result<IAuthenticationModel, Error>
-	func signInSocial(_ socialType: SocialType, domain: String)
+	func signInSocial(_ socialType: SocialType, domain: String) async -> Result<IAuthenticationModel, Error>
 }
 
 struct LoginRemoteStore {
@@ -32,14 +33,25 @@ extension LoginRemoteStore: ILoginRemoteStore {
 		}
 	}
 	
-	func signInSocial(_ socialType: SocialType, domain: String) {
+	func signInSocial(_ socialType: SocialType, domain: String) async -> Result<IAuthenticationModel, Error> {
+		var result: Result<Auth_SocialLoginRes, Error>?
+		
 		switch socialType {
 		case .facebook:
-			socialAuthenticationService.signInWithFB(domain: domain)
+			result = await socialAuthenticationService.signInWithFB(domain: domain)
 		case .google:
-			socialAuthenticationService.signInWithGoogle(domain: domain)
+			result = await socialAuthenticationService.signInWithGoogle(domain: domain)
 		case .office:
-			socialAuthenticationService.signInWithOffice(domain: domain)
+			result = await socialAuthenticationService.signInWithOffice(domain: domain)
+		}
+		
+		switch result {
+		case .success(let socialLoginResponse):
+			return .success(AuthenticationModel(response: socialLoginResponse))
+		case .failure(let error):
+			return .failure(error)
+		case .none:
+			return .failure(ServerError.unknown)
 		}
 	}
 }
