--- Packages/CommonUI/CommonUI/Classes/View/Extension/Extension.swift
@@ -1,16 +1,0 @@
-//
-//  Extension.swift
-//  _NIODataStructures
-//
-//  Created by MinhDev on 26/04/2022.
-//
-
-import SwiftUI
-
-#if canImport(UIKit)
-public extension View {
-	func hideKeyboard() {
-		UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-	}
-}
-#endif
