+++ Packages/CommonUI/CommonUI/Classes/View/Extension/View+Extension.swift
@@ -1,0 +1,35 @@
+//
+//  Extension.swift
+//  _NIODataStructures
+//
+//  Created by MinhDev on 26/04/2022.
+//
+
+import SwiftUI
+
+#if canImport(UIKit)
+public extension View {
+	func hideKeyboard() {
+		UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+	}
+}
+#endif
+
+public extension View {
+	func hiddenNavigationBarStyle() -> some View {
+		modifier( HiddenNavigationBar() )
+	}
+	
+	/// Applies the given transform if the given condition evaluates to `true`.
+	/// - Parameters:
+	///   - condition: The condition to evaluate.
+	///   - transform: The transform to apply to the source `View`.
+	/// - Returns: Either the original `View` or the modified `View` if the condition is `true`.
+	@ViewBuilder func `if`<Content: View>(_ condition: Bool, transform: (Self) -> Content) -> some View {
+		if condition {
+			transform(self)
+		} else {
+			self
+		}
+	}
+}
